+++ cli/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased
 ----------
 - Added `--user-agent` option for configuring the user agent to use
+- Added `--remove-selector` option for removing elements before a
+  screenshot is taken
 
 
 0.1.0